DronesSimulation become a real thing

+++ src/show_simulation/convex_hull.py
@@ -30,19 +30,20 @@ def two_dimensionnal_cross_product(
     ) * (position_2[0] - position_0[0])
 
 
-def tuple_list_to_array_list(tuple_list: List[Tuple[int, int]]) -> List[np.ndarray]:
+def tuple_list_to_array_list(tuple_list: List[Tuple[float, float]]) -> List[np.ndarray]:
     return [np.array(tuple_element) for tuple_element in tuple_list]
 
 
-def array_list_to_tuple_list(array_list: List[np.ndarray]) -> List[Tuple[int, int]]:
+def array_list_to_tuple_list(array_list: List[np.ndarray]) -> List[Tuple[float, float]]:
     return [
-        (int(array_element[0]), int(array_element[1])) for array_element in array_list
+        (float(array_element[0]), float(array_element[1]))
+        for array_element in array_list
     ]
 
 
 def calculate_convex_hull(
-    positions_tuple: List[Tuple[int, int]]
-) -> List[Tuple[int, int]]:
+    positions_tuple: List[Tuple[float, float]]
+) -> List[Tuple[float, float]]:
     """Graham scan implementation"""
     positions_array = tuple_list_to_array_list(positions_tuple)
